feat: 랜딩페이지 애니메이션 적용 완료
