+++ src/feature/landing/components/TodayWineSection.tsx
@@ -1,5 +1,79 @@
+'use client';
+import Image from 'next/image';
+import { useScroll, useTransform, useInView } from 'motion/react';
+import * as motion from 'motion/react-client';
+import { useRef } from 'react';
+import {
+  section2TitleVariants,
+  section2SubTitleVariants,
+} from '../utils/animations';
+import { STYLES } from '../constants/styles';
+
 const TodayWineSection = () => {
-  return <div>TodayWineSection</div>;
+  const section2Ref = useRef(null);
+  const { scrollYProgress } = useScroll({
+    target: section2Ref,
+    offset: ['start end', 'end start'],
+  });
+
+  const redWineRotate = useTransform(scrollYProgress, [0, 1], [-10, 5]);
+  const whiteWineRotate = useTransform(scrollYProgress, [0, 1], [-30, 5]);
+  const isInView2 = useInView(section2Ref, {
+    once: true,
+    margin: '0px 0px -30% 0px',
+  });
+
+  return (
+    <section
+      ref={section2Ref}
+      className='relative mx-auto mt-[10rem] flex h-[50.8rem] max-w-[34.3rem] justify-center pt-[13rem]'
+    >
+      <motion.div
+        style={{ rotate: redWineRotate }}
+        className='absolute top-0 left-0 z-[-1] origin-bottom'
+      >
+        <Image
+          src='/images/landing/image-wine-red-landing.png'
+          alt='wine-red'
+          width={207}
+          height={367}
+        />
+      </motion.div>
+
+      <motion.div
+        style={{ rotate: whiteWineRotate }}
+        className='absolute bottom-0 left-[-2.3rem] z-[-2] max-w-none origin-top'
+      >
+        <Image
+          src='/images/landing/image-wine-white-landing.png'
+          alt='wine-white'
+          width={375}
+          height={364}
+        />
+      </motion.div>
+
+      <div className='w-[16.5rem] overflow-hidden'>
+        <motion.h2
+          className={STYLES.sectionTitle}
+          variants={section2TitleVariants}
+          initial='hidden'
+          animate={isInView2 ? 'visible' : 'hidden'}
+        >
+          TODAY <br /> WINE
+        </motion.h2>
+        <motion.p
+          className={STYLES.sectionSubTitle}
+          variants={section2SubTitleVariants}
+          initial='hidden'
+          animate={isInView2 ? 'visible' : 'hidden'}
+        >
+          매달 새롭게 추천되는 <br />
+          다양한 인기 와인을 <br />
+          만나보세요.
+        </motion.p>
+      </div>
+    </section>
+  );
 };
 
 export default TodayWineSection;
