+++ src/feature/landing/constants/content.ts
@@ -1,0 +1,31 @@
+export const HERO_TEXT_LINES = ['CODEIT SPRINT FRONTEND', 'PART3 TEAM7'];
+
+export const HERO_PARAGRAPH_LINES = [
+  'This is a rebranding of "WHYNE"',
+  'from the intermediate project theme.',
+  'The vintage beige colour is',
+  'used to complement',
+  'the warm red colour of wine,',
+  'and we want to give users a',
+  'warm wine',
+  'research experience.',
+];
+
+export const TEAM_MEMBERS = [
+  'JAEHYUN',
+  'SUMIN',
+  'YUNGHYUN',
+  'JINHYUNG',
+  'SARANG',
+];
+
+export const NOTE_IMAGES = [
+  'cherry',
+  'peach',
+  'oak',
+  'star-anise',
+  'tobacco',
+  'herb',
+  'vanilla',
+  'apple',
+];
