+++ src/feature/landing/constants/styles.ts
@@ -1,0 +1,7 @@
+export const STYLES = {
+  listItem:
+    'border-primary text-primary text-medium flex h-[3.1rem] w-[7.4rem] items-center justify-center rounded-[50%] border-1 text-[1.3rem] font-medium tracking-[-0.065rem]',
+  sectionTitle: 'text-primary text-[4.8rem] font-extrabold leading-[4rem]',
+  sectionSubTitle:
+    'text-primary text-[1.4rem] font-light mt-[0.4rem] leading-[1.8rem]',
+} as const;
