+++ src/feature/landing/utils/text.ts
@@ -1,0 +1,6 @@
+export function splitText(text: string) {
+  return text.split('').map((char, index) => ({
+    char,
+    key: `${char}-${index}`,
+  }));
+}
